split useConnectWorldUsers into it's own file
